♻️ refactor(api/common/schemas): remove FlattenUnion function and update imports in schemas
