+++ apps/api/src/modules/with-client-type/with-client-type.plugin.ts
@@ -2,7 +2,8 @@ import { Value } from "@sinclair/typebox/value";
 import { Elysia, t } from "elysia";
 import { CLIENT_TYPE_HEADER_NAME } from "../../common/constants";
 import { type ClientType, clientTypeSchema, newClientType } from "../../domain/value-object";
-import { BadRequestException, ErrorResponseSchema } from "../error";
+import { ErrorResponseSchema } from "../elysia-with-env";
+import { BadRequestException } from "../error";
 
 export const withClientType = new Elysia({
 	name: "@mona-ca/with-client-type",
